// src/shared/types/Genre.ts
enum Genre{
    ACTION = "Action",
    COMEDY = "Comedy",
    ADVENTURE = "Adventure",
    ANIMATION = "Animation",
    CRIME = "Crime",
    DOCUMENTRY = "Documentry",
    DRAMA = "Drama",
    FAMILY = "Family",
    FANTASY = "Fantasy",
    HISTORY = "History",
    HORROR = "Horror",
    MUSIC = "Music",
    MYSTERY = "Mystery",
    ROMANCE = "Romance",
    SCI_FI = "Sci-fi",
    THRILLER = "Thriller",
    WAR = "War",
    WESTERN = "Western",
    TV = "Tv"
}

export default Genre;